address reviewed changes
